requirements.txt
tensorflow==1.12.0
flask==1.0.2
requests==2.21.0
Pillow==5.4.1
pyrebase=3.0.27
